Fix crash caused by falling nodes
The falling nodes code iterates through and calls
on_dignode callbacks when a falling entities lands
on a node that is not a full block (such as a torch)
and removes it. However, no player is provided to these
callback functions. In order to compensate for this,
whether or not the player variable is defined is first
checked before attempting to call any other gamemode
logic, and if it does not exist, the event is ignored
and allowed to continue as intended.